Decouple Client HTML from Video Asset
Before the client HTML was heavily coupled to the video asset
that was being served. This required the HTML to be generated
each time.

In this change, the JavaScript that starts playback has been
changed to read the asset resource from the url params. This
allows the same client code to be used regardless of asset.

As part of this decoupling, the HTML, CSS, and JavaScript have
been broken-up and moved to a client directory. This was to help
them be viewable in isolation.